\name{denboer2009.group.colors}
\alias{denboer2009.group.colors}
\docType{data}
\title{Group-specific colors for the denboer2009 dataset}
\description{
Group-specific colors to highlight subtypes of ALL in drawings for the denboer2009 dataset
}
\usage{data("denboer2009.group.colors")}
\format{A named vector associating one color (value) to each subtype of ALL (name).}
\references{
Den Boer, M.L., van Slegtenhorst, M., De Menezes, R.X., Cheok, M.H., Buijs-Gladdines, J.G.C.A.M., Peters, S.T.C.J.M., Van Zutven, L.J.C.M., Beverloo, H.B., Van der Spek, P.J., Escherich, G., et al. (2009) A subtype of childhood acute lymphoblastic leukaemia with poor treatment outcome: a genome-wide classification study. Lancet Oncol, 10, 125–134.
}
\examples{
data(denboer2009.group.colors)

## Display the group-color associations
data.frame(denboer2009.group.colors)
}
\keyword{datasets}
